Fixed compilation issues
git-svn-id: http://miranda.googlecode.com/svn/trunk@9150 83df23c2-51fd-11de-bb24-2d5e9b4af41d

+++ miranda/protocols/MSN/msn_chat.cpp
@@ -23,7 +23,7 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
 #include "msn_proto.h"
 #include <m_history.h>
 
-static LONG sttChatID = 0;
+static INT sttChatID = 0;
 
 HANDLE CMsnProto::MSN_GetChatInernalHandle(HANDLE hContact)
 {
